Revert "Add Import Policy for Service VIPs (#721)"
This reverts commit 8fe9f70dd59a7fc5c0de7ffcd07b75e7b0b99a8c.
