+++ pkg/controllers/routing/network_routes_controller.go
@@ -286,9 +286,9 @@ func (nrc *NetworkRoutingController) Run(healthChan chan<- *healthcheck.Controll
 			glog.Errorf("Error advertising route: %s", err.Error())
 		}
 
-		err = nrc.AddPolicies()
+		err = nrc.addExportPolicies()
 		if err != nil {
-			glog.Errorf("Error adding BGP policies: %s", err.Error())
+			glog.Errorf("Error adding BGP export policies: %s", err.Error())
 		}
 
 		if nrc.bgpEnableInternal {
